Fix RUBY_GEMS_ROOT
With the addition from ea8c9653:

    Install ruby with --system

To the Dockerfile, this makes it so ruby is no longer installed into
`/opt/rubies/ruby-2.3.1`, but instead `/usr/local/lib/ruby` (the
behavior of `--system` in `ruby-install`).  This fix updates the
`RUBY_GEMS_ROOT` ENV variable to reflect that properly.

This also means that `/usr/local/bin` already has the `ruby` binary
installed, so modifying the `$PATH` is no longer necessary.